+++ game/settings/settings.py
@@ -443,6 +443,27 @@ class Settings:
         section=PERFORMANCE_SECTION,
         default=True,
     )
+    # perf_disable_convoys: bool = boolean_option(
+    #     "Disable convoys",
+    #     page=MISSION_GENERATOR_PAGE,
+    #     section=PERFORMANCE_SECTION,
+    #     default=False,
+    # )
+    # perf_frontline_units_prefer_roads: bool = boolean_option(
+    #     "Front line troops prefer roads",
+    #     page=MISSION_GENERATOR_PAGE,
+    #     section=PERFORMANCE_SECTION,
+    #     default=False,
+    # )
+    perf_frontline_units_max_supply: int = bounded_int_option(
+        "Maximum frontline unit supply per control point",
+        page=MISSION_GENERATOR_PAGE,
+        section=PERFORMANCE_SECTION,
+        default=60,
+        min=10,
+        max=300,
+        causes_expensive_game_update=True,
+    )
     perf_infantry: bool = boolean_option(
         "Generate infantry squads alongside vehicles",
         page=MISSION_GENERATOR_PAGE,